# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMS Hub API",
        "description": "SMS Hub API Documentation - A comprehensive SMS gateway service with multiple provider support",
        "contact": {
            "name": "SMS Hub Support",
            "email": "support@smshub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/webhooks/{provider}/dlr": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Delivery report webhook",
                "description": "Receive delivery reports from SMS providers",
                "operationId": "deliveryReport",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "SMS provider name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "eskiz",
                                "playmobile"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_id",
                                    "status"
                                ],
                                "properties": {
                                    "message_id": {
                                        "description": "Message ID",
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "status": {
                                        "description": "Delivery status",
                                        "type": "string",
                                        "example": "delivered"
                                    },
                                    "error_code": {
                                        "description": "Error code if failed",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "error_message": {
                                        "description": "Error message if failed",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "delivered_at": {
                                        "description": "Delivery timestamp",
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-01-15T10:30:00Z"
                                    },
                                    "cost": {
                                        "description": "Message cost",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.05
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Delivery report processed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid webhook data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/webhooks/{provider}/status": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Status update webhook",
                "description": "Receive status updates from SMS providers",
                "operationId": "statusUpdate",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "SMS provider name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "eskiz",
                                "playmobile"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_id",
                                    "status"
                                ],
                                "properties": {
                                    "message_id": {
                                        "description": "Message ID",
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "status": {
                                        "description": "Message status",
                                        "type": "string",
                                        "example": "sent"
                                    },
                                    "provider_message_id": {
                                        "description": "Provider's message ID",
                                        "type": "string",
                                        "example": "provider_12345"
                                    },
                                    "updated_at": {
                                        "description": "Status update timestamp",
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-01-15T10:30:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status update processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Status update processed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid webhook data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response",
                "properties": {
                    "error": {
                        "description": "Error type",
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Invalid or missing authentication token"
                    },
                    "code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "scopes": {
                            "read": "Read access",
                            "write": "Write access",
                            "admin": "Admin access"
                        }
                    },
                    "password": {
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "scopes": {
                            "read": "Read access",
                            "write": "Write access",
                            "admin": "Admin access"
                        }
                    }
                }
            },
            "bearer": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer": []
        }
    ]
}